Fix MissingMethodException on mod menu creation
Due to a Satchel update the SetMenuButtonNameAndDesc function no longer
exists, this causes a `MissingMethodException` when the mapi tries to
create the mod menu. But since the method in this case is only used to
change the name of the button and not description the behaviour can be
replicated by implementing a custom GetMenuButtonText method.